Starts modifying display text when looking at interactive. Modifies sprint.
After talking with a few Columbia students, each of them agreed that they prefer a hold to sprint instead of a toggle. And relooking at the game and seeing that the player won't be sprinting across huge worlds, the decision was made to make it hold. Added a bool in case I want to change it later.