Adds audio sources to all itneracrtable objects
The base Interactable class will be made abstract and be sublassed into different items. Every sublass will ahve an audio source component and play a sound effect when being picked up. Ammo will sound like a box of bullets being shaken, a readable object will play a page flipping sfx, and so on. I was origianlly hesitant about making the decision to make every interactable onject have a sound effect, as I felt like thert would be some objects to bei teracted with that would not need SFX, but I came to disagree strongly on that position. ONeo f the greatest mistakes is lack of feedback in games. when developing, especialyl as a student or as someoen who's inexpereinced, you shoudl try designing feedback to be dailed to 11, and then some more. Try your best to go overboard, and then some.